Refactor: moved download function to a helper for re-useability in future
